requirements.txt
anyio==4.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
numpy==1.26.4
pandas==2.2.0
PyMySQL==1.1.0
python-telegram-bot==20.7
pytz==2024.1
requests==2.31.0
sniffio==1.3.0
soupsieve==2.5
tzdata==2024.1
urllib3==2.2.0
